Small Screen Alternating Section Background Fix
Since <main> now has an even number of sections, tweaked the selector on lines 192 and 193 of main.css to preserve the alternating background pattern on small screens between the #publications and #bioelectricity-seminars sections. 

The change in this commit would need to be undone if <main> went back to having an odd number of sections (my apologies, it is a quirk of using alternating backgrounds with a two column layout).